+++ src/io/github/frenchy64/fully_satisfies/everyp.clj
@@ -18,45 +18,45 @@
     (defn everyp [& ps]
       (fn [& args] (every? #(every? % args) ps)))"
   {:arglists '([& ps])}
-  ([] ;; fully-satisfies: added zero-arity
-     (fn
-       ([] true)
-       ([x] true)
-       ([x y] true)
-       ([x y z] true)
-       ([x y z & args] true)))
-  ([p]
-     (fn ep1
-       ([] true)
-       ([x] (boolean (p x)))
-       ([x y] (boolean (and (p x) (p y))))
-       ([x y z] (boolean (and (p x) (p y) (p z))))
-       ([x y z & args] (boolean (and (ep1 x y z)
-                                     (every? p args))))))
-  ([p1 p2]
-     (fn ep2
-       ([] true)
-       ([x] (boolean (and (p1 x) (p2 x))))
-       ([x y] (boolean (and (p1 x) (p1 y) (p2 x) (p2 y))))
-       ([x y z] (boolean (and (p1 x) (p1 y) (p1 z) (p2 x) (p2 y) (p2 z))))
-       ([x y z & args] (boolean (and (ep2 x y z)
-                                     (every? #(and (p1 %) (p2 %)) args))))))
-  ([p1 p2 p3]
-     (fn ep3
-       ([] true)
-       ([x] (boolean (and (p1 x) (p2 x) (p3 x))))
-       ([x y] (boolean (and (p1 x) (p1 y) (p2 x) (p2 y) (p3 x) (p3 y))))
-       ([x y z] (boolean (and (p1 x) (p1 y) (p1 z) (p2 x) (p2 y) (p2 z) (p3 x) (p3 y) (p3 z))))
-       ([x y z & args] (boolean (and (ep3 x y z)
-                                     (every? #(and (p1 %) (p2 %) (p3 %)) args))))))
-  ([p1 p2 p3 & ps]
-     (let [ps (list* p1 p2 p3 ps)]
-       (fn epn
-         ([] true)
-         ([x] (every? #(% x) ps))
-         ([x y] (every? #(and (% x) (% y)) ps))
-         ([x y z] (every? #(and (% x) (% y) (% z)) ps))
-         ([x y z & args] (boolean (and (epn x y z)
-                                       (every? #(every? % args) ps))))))))
+  ([] (fn
+        ([] true)
+        ([x] true)
+        ([x y] true)
+        ([x y z] true)
+        ([x y z & args] true)))
+  ([p1] (fn
+          ([] true)
+          ([x] (boolean (p1 x)))
+          ([x y] (boolean (and (p1 x) (p1 y))))
+          ([x y z] (boolean (and (p1 x) (p1 y) (p1 z))))
+          ([x y z & args] (boolean (and (p1 x) (p1 y) (p1 z) (every? p1 args))))))
+  ([p1 p2] (fn
+             ([] true)
+             ([x] (boolean (and (p1 x) (p2 x))))
+             ([x y] (let [tp (fn [p] (and (p x) (p y)))]
+                      (boolean (and (tp p1) (tp p2)))))
+             ([x y z] (let [tp (fn [p] (and (p x) (p y) (p z)))]
+                        (boolean (and (tp p1) (tp p2)))))
+             ([x y z & args] (let [tp (fn [p] (and (p x) (p y) (p z) (every? p args)))]
+                               (boolean (and (tp p1) (tp p2)))))))
+  ([p1 p2 p3] (fn
+                ([] true)
+                ([x] (boolean (and (p1 x) (p2 x) (p3 x))))
+                ([x y] (let [tp (fn [p] (and (p x) (p y)))]
+                         (boolean (and (tp p1) (tp p2) (tp p3)))))
+                ([x y z] (let [tp (fn [p] (and (p x) (p y) (p z)))]
+                           (boolean (and (tp p1) (tp p2) (tp p3)))))
+                ([x y z & args] (let [tp (fn [p] (and (p x) (p y) (p z) (every? p args)))]
+                                  (boolean (and (tp p1) (tp p2) (tp p3)))))))
+  ([p1 p2 p3 & ps] (fn
+                     ([] true)
+                     ([x] (boolean (and (p1 x) (p2 x) (p3 x) (every? (fn [p] (p x)) ps))))
+                     ([x y] (let [tp (fn [p] (and (p x) (p y)))]
+                              (boolean (and (tp p1) (tp p2) (tp p3) (every? tp ps)))))
+                     ([x y z] (let [tp (fn [p] (and (p x) (p y) (p z)))]
+                                (boolean (and (tp p1) (tp p2) (tp p3) (every? tp ps)))))
+                     ([x y z & args] (let [tp (fn [p] (and (p x) (p y) (p z) (every? p args)))]
+                                       (boolean (and (tp p1) (tp p2) (tp p3) (every? tp ps))))))))
+
 
 (def every-pred everyp)