+++ test/io/github/frenchy64/fully_satisfies/everyp_test.clj
@@ -1,6 +1,7 @@
 (ns io.github.frenchy64.fully-satisfies.everyp-test
   (:refer-clojure :exclude [every-pred])
   (:require [clojure.test :refer [is]]
+            [clojure.math.combinatorics :as comb]
             [com.gfredericks.test.chuck.clojure-test :refer [checking]]
             [io.github.frenchy64.fully-satisfies.never :refer [never?]]
             [io.github.frenchy64.fully-satisfies.uncaught-testing-contexts :refer [testing deftest]]
@@ -11,8 +12,19 @@
 
 ;; TODO order of operations
 ;; TODO test zero arity
-#_ ;;FIXME
 (deftest everyp-test
+  (doseq [i (range 7)
+          false-args (map set (comb/subsets (range i)))
+          :let [args (map (fn [i] (if (false-args i) true false))
+                          (range i))]
+          false-preds (map set (comb/subsets (range i)))
+          :let [preds (map (fn [i] (if (false-preds i) true? false?))
+                           (range i))]]
+    (is (= (apply (apply everyp preds) args)
+           (apply (apply every-pred preds) args)
+           (apply (apply everyp-reference preds) args))
+        [args preds]))
+#_ ;;FIXME
   (doseq [everyp [everyp every-pred everyp-reference]]
     (testing everyp
       (testing "found match"